requirements.txt
torch==2.3.0
torchvision==0.18.0
opencv-python==4.9.0.80
numpy==1.26.4
pandas==2.2.2
matplotlib==3.8.4
seaborn==0.13.2
tensorboard==2.16.2
pyyaml==6.0.1
tqdm==4.66.4
imageio==2.34.0
opencv-python-headless==4.9.0.80  # Replacing existing opencv-python with headless version